# Improve ROS 2 builtin type `string` handling