--- a/lib/rclex/generators/msg_ex.ex
+++ b/lib/rclex/generators/msg_ex.ex
@@ -138,7 +138,7 @@ defmodule Rclex.Generators.MsgEx do
         # credo:disable-for-next-line Credo.Check.Refactor.Nesting
         case [type_tuple, name] do
           [{:builtin_type, "string" <> _}, name] ->
-            ~s/~c"\#{#{name}}"/
+            "#{name}"
 
           [{:builtin_type, _type}, name] ->
             "#{name}"
@@ -181,7 +181,7 @@ defmodule Rclex.Generators.MsgEx do
         # credo:disable-for-next-line Credo.Check.Refactor.Nesting
         case [type_tuple, name] do
           [{:builtin_type, "string" <> _}, name] ->
-            ~s/#{name}: "\#{#{name}}"/
+            "#{name}: #{name}"
 
           [{:builtin_type, _type}, name] ->
             "#{name}: #{name}"