--- a/scripts/README.md
+++ b/scripts/README.md
@@ -1,0 +1,13 @@
+# About scripts
+
+These scripts are for development. So package doesn't include these.
+
+### `*_benchee.exs`
+
+You can run them by `mix run scripts/*_benchee.exs` under the project.
+
+### `iwyu.sh`
+
+You can use `iwyu`, `include-what-you-use` by `mix iwyu` which call this script.
+
+NOTE: `iwyu` is `[include-what-you-use](https://github.com/include-what-you-use/include-what-you-use)`.