--- a/test/expected_files/sensor_msgs/msg/point_cloud_get_fun.txt
+++ b/test/expected_files/sensor_msgs/msg/point_cloud_get_fun.txt
@@ -21,7 +21,7 @@
     }
 
     channels[channels_i] = enif_make_tuple(env, 2,
-      enif_make_string(env, message_p->channels.data[channels_i].name.data, ERL_NIF_LATIN1),
+      enif_make_binary_wrapper(env, message_p->channels.data[channels_i].name.data, message_p->channels.data[channels_i].name.size),
       enif_make_list_from_array(env, channels_values, message_p->channels.data[channels_i].values.size)
     );
   }
@@ -32,7 +32,7 @@
         enif_make_int(env, message_p->header.stamp.sec),
         enif_make_uint(env, message_p->header.stamp.nanosec)
       ),
-      enif_make_string(env, message_p->header.frame_id.data, ERL_NIF_LATIN1)
+      enif_make_binary_wrapper(env, message_p->header.frame_id.data, message_p->header.frame_id.size)
     ),
     enif_make_list_from_array(env, points, message_p->points.size),
     enif_make_list_from_array(env, channels, message_p->channels.size)