--- a/test/expected_files/sensor_msgs/msg/point_cloud_set_fun.txt
+++ b/test/expected_files/sensor_msgs/msg/point_cloud_set_fun.txt
@@ -18,20 +18,11 @@
     return enif_make_badarg(env);
   message_p->header.stamp.nanosec = header_stamp_nanosec;
 
-  unsigned int header_frame_id_length;
-#if (ERL_NIF_MAJOR_VERSION == 2 && ERL_NIF_MINOR_VERSION >= 17) // OTP-26 and later
-  if (!enif_get_string_length(env, header_tuple[1], &header_frame_id_length, ERL_NIF_LATIN1))
-    return enif_make_badarg(env);
-#else
-  if (!enif_get_list_length(env, header_tuple[1], &header_frame_id_length))
-    return enif_make_badarg(env);
-#endif
-
-  char header_frame_id[header_frame_id_length + 1];
-  if (enif_get_string(env, header_tuple[1], header_frame_id, header_frame_id_length + 1, ERL_NIF_LATIN1) <= 0)
+  ErlNifBinary header_frame_id_binary;
+  if (!enif_inspect_binary(env, header_tuple[1], &header_frame_id_binary))
     return enif_make_badarg(env);
 
-  if (!rosidl_runtime_c__String__assign(&(message_p->header.frame_id), header_frame_id))
+  if (!rosidl_runtime_c__String__assignn(&(message_p->header.frame_id), (const char *)header_frame_id_binary.data, header_frame_id_binary.size))
     return raise(env, __FILE__, __LINE__);
 
   unsigned int points_length;
@@ -90,20 +81,11 @@
     if (!enif_get_tuple(env, channels_head, &channels_i_arity, &channels_i_tuple))
       return enif_make_badarg(env);
 
-    unsigned int channels_i_name_length;
-#if (ERL_NIF_MAJOR_VERSION == 2 && ERL_NIF_MINOR_VERSION >= 17) // OTP-26 and later
-    if (!enif_get_string_length(env, channels_i_tuple[0], &channels_i_name_length, ERL_NIF_LATIN1))
-      return enif_make_badarg(env);
-#else
-    if (!enif_get_list_length(env, channels_i_tuple[0], &channels_i_name_length))
-      return enif_make_badarg(env);
-#endif
-
-    char channels_i_name[channels_i_name_length + 1];
-    if (enif_get_string(env, channels_i_tuple[0], channels_i_name, channels_i_name_length + 1, ERL_NIF_LATIN1) <= 0)
+    ErlNifBinary channels_i_name_binary;
+    if (!enif_inspect_binary(env, channels_i_tuple[0], &channels_i_name_binary))
       return enif_make_badarg(env);
 
-    if (!rosidl_runtime_c__String__assign(&(message_p->channels.data[channels_i].name), channels_i_name))
+    if (!rosidl_runtime_c__String__assignn(&(message_p->channels.data[channels_i].name), (const char *)channels_i_name_binary.data, channels_i_name_binary.size))
       return raise(env, __FILE__, __LINE__);
 
     unsigned int channels_i_values_length;