--- a/test/expected_files/std_msgs/msg/multi_array_dimension_get_fun.txt
+++ b/test/expected_files/std_msgs/msg/multi_array_dimension_get_fun.txt
@@ -1,5 +1,5 @@
   return enif_make_tuple(env, 3,
-    enif_make_string(env, message_p->label.data, ERL_NIF_LATIN1),
+    enif_make_binary_wrapper(env, message_p->label.data, message_p->label.size),
     enif_make_uint(env, message_p->size),
     enif_make_uint(env, message_p->stride)
   );