--- a/test/expected_files/std_msgs/msg/multi_array_dimension_set_fun.txt
+++ b/test/expected_files/std_msgs/msg/multi_array_dimension_set_fun.txt
@@ -1,17 +1,8 @@
-  unsigned int label_length;
-#if (ERL_NIF_MAJOR_VERSION == 2 && ERL_NIF_MINOR_VERSION >= 17) // OTP-26 and later
-  if (!enif_get_string_length(env, tuple[0], &label_length, ERL_NIF_LATIN1))
-    return enif_make_badarg(env);
-#else
-  if (!enif_get_list_length(env, tuple[0], &label_length))
-    return enif_make_badarg(env);
-#endif
-
-  char label[label_length + 1];
-  if (enif_get_string(env, tuple[0], label, label_length + 1, ERL_NIF_LATIN1) <= 0)
+  ErlNifBinary label_binary;
+  if (!enif_inspect_binary(env, tuple[0], &label_binary))
     return enif_make_badarg(env);
 
-  if (!rosidl_runtime_c__String__assign(&(message_p->label), label))
+  if (!rosidl_runtime_c__String__assignn(&(message_p->label), (const char *)label_binary.data, label_binary.size))
     return raise(env, __FILE__, __LINE__);
 
   unsigned int size;