--- a/test/expected_files/std_msgs/msg/string.c
+++ b/test/expected_files/std_msgs/msg/string.c
@@ -76,20 +76,11 @@ ERL_NIF_TERM nif_std_msgs_msg_string_set(ErlNifEnv *env, int argc, const ERL_NIF
   const ERL_NIF_TERM *tuple;
   if (!enif_get_tuple(env, argv[1], &arity, &tuple)) return enif_make_badarg(env);
 
-  unsigned int data_length;
-#if (ERL_NIF_MAJOR_VERSION == 2 && ERL_NIF_MINOR_VERSION >= 17) // OTP-26 and later
-  if (!enif_get_string_length(env, tuple[0], &data_length, ERL_NIF_LATIN1))
-    return enif_make_badarg(env);
-#else
-  if (!enif_get_list_length(env, tuple[0], &data_length))
-    return enif_make_badarg(env);
-#endif
-
-  char data[data_length + 1];
-  if (enif_get_string(env, tuple[0], data, data_length + 1, ERL_NIF_LATIN1) <= 0)
+  ErlNifBinary data_binary;
+  if (!enif_inspect_binary(env, tuple[0], &data_binary))
     return enif_make_badarg(env);
 
-  if (!rosidl_runtime_c__String__assign(&(message_p->data), data))
+  if (!rosidl_runtime_c__String__assignn(&(message_p->data), (const char *)data_binary.data, data_binary.size))
     return raise(env, __FILE__, __LINE__);
 
   return atom_ok;
@@ -105,7 +96,7 @@ ERL_NIF_TERM nif_std_msgs_msg_string_get(ErlNifEnv *env, int argc, const ERL_NIF
   std_msgs__msg__String *message_p = (std_msgs__msg__String *)*ros_message_pp;
 
   return enif_make_tuple(env, 1,
-    enif_make_string(env, message_p->data.data, ERL_NIF_LATIN1)
+    enif_make_binary_wrapper(env, message_p->data.data, message_p->data.size)
   );
 }
 // clang-format on