--- a/test/expected_files/std_msgs/msg/string_get_fun.txt
+++ b/test/expected_files/std_msgs/msg/string_get_fun.txt
@@ -1,3 +1,3 @@
   return enif_make_tuple(env, 1,
-    enif_make_string(env, message_p->data.data, ERL_NIF_LATIN1)
+    enif_make_binary_wrapper(env, message_p->data.data, message_p->data.size)
   );