--- a/test/expected_files/std_msgs/msg/string_set_fun.txt
+++ b/test/expected_files/std_msgs/msg/string_set_fun.txt
@@ -1,15 +1,6 @@
-  unsigned int data_length;
-#if (ERL_NIF_MAJOR_VERSION == 2 && ERL_NIF_MINOR_VERSION >= 17) // OTP-26 and later
-  if (!enif_get_string_length(env, tuple[0], &data_length, ERL_NIF_LATIN1))
-    return enif_make_badarg(env);
-#else
-  if (!enif_get_list_length(env, tuple[0], &data_length))
-    return enif_make_badarg(env);
-#endif
-
-  char data[data_length + 1];
-  if (enif_get_string(env, tuple[0], data, data_length + 1, ERL_NIF_LATIN1) <= 0)
+  ErlNifBinary data_binary;
+  if (!enif_inspect_binary(env, tuple[0], &data_binary))
     return enif_make_badarg(env);
 
-  if (!rosidl_runtime_c__String__assign(&(message_p->data), data))
+  if (!rosidl_runtime_c__String__assignn(&(message_p->data), (const char *)data_binary.data, data_binary.size))
     return raise(env, __FILE__, __LINE__);