… test/expected_files/std_msgs/msg/multi_array_layout_set_fun.txt
… test/expected_files/std_msgs/msg/u_int32_multi_array_get_fun.txt
… test/expected_files/std_msgs/msg/u_int32_multi_array_set_fun.txt