This PR improves ROS 2 builtin type `string` handling, `get!` and `set!`, especially for long string message, like over `2**16` bytes.

Until now we treat the `string` as Erlang string, it means char array, this PR changes it as Erlang binary.
This PR doesn't change user API.

Left todos are following,

- [x] fix tests

## bench

### env

```
Operating System: Linux
CPU Information: Intel(R) Core(TM) Ultra 7 155U
Number of Available Cores: 14
Available memory: 30.86 GB
Elixir 1.17.3
Erlang 27.2.4
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 0 ns
reduction time: 0 ns
parallel: 1
inputs: none specified
Estimated total run time: 28 s
```

### Before

```bash
Name                 ips        average  deviation         median         99th %
create!/0      1200.10 K        0.83 μs  ±7282.50%       0.155 μs        0.56 μs
destroy!/1      613.62 K        1.63 μs   ±197.89%        1.48 μs        3.91 μs
get!/1            2.69 K      371.08 μs    ±50.20%      259.41 μs      924.29 μs
set!/2            2.59 K      386.75 μs    ±50.62%      267.64 μs      904.49 μs

Comparison:
create!/0      1200.10 K
destroy!/1      613.62 K - 1.96x slower +0.80 μs
get!/1            2.69 K - 445.33x slower +370.25 μs
set!/2            2.59 K - 464.13x slower +385.91 μs
```

### After

```bash
Name                 ips        average  deviation         median         99th %
create!/0      1306.41 K        0.77 μs  ±7568.95%        0.22 μs        0.33 μs
destroy!/1      846.92 K        1.18 μs    ±24.47%        1.15 μs        1.46 μs
set!/2          404.17 K        2.47 μs   ±363.48%        2.39 μs        3.35 μs
get!/1          366.45 K        2.73 μs   ±360.36%        2.49 μs        4.92 μs

Comparison:
create!/0      1306.41 K
destroy!/1      846.92 K - 1.54x slower +0.42 μs
set!/2          404.17 K - 3.23x slower +1.71 μs
get!/1          366.45 K - 3.57x slower +1.96 μs
```