@pojiro 2025-03-07
@takasehideki Thanks for paying attention to this PR. Could you review this?
I'd like to compare performance between rclex and rclcpp again with this improvement,  on RPI5 with jazzy.