@pojiro 2025-03-08
@takasehideki 

> My poor understanding is that your proposal is to treat string properly as binary in Erlang to be faster even for large sizes.

https://github.com/rclex/rclex/blob/72afa507ef9232794362ab1b9a6f180c0775b19f/lib/rclex/generators/msg_c.ex#L361-L363

There are two improvements.

1. quit to use VLA ([Variable-length array](https://en.wikipedia.org/wiki/Variable-length_array#:~:text=In%20computer%20programming%2C%20a%20variable,value%20(see%20Dependent%20type).)). It was used for the line 361. VLA uses stack, so using large length string may cause stack overflow. Actually 2^^20 bytes string cause it at my PC.
2. quit to treat as `string`. This PR change it from string to binary. This means Erlang/Elixir don't care about the content's encodings. So this makes thing faster. This also remove the overhead of conversion between Erlang charlist and binary. This makes thing faster too.

> there were a few things I didn't understand, so I made some comments. These are just trivial questions

I'd like to resolve your question, so feel free to ask anything about my PR.